# ERP/docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: ERP Main APIs for Apifox
  version: "1.0.0"
  description: |
    Main APIs extracted from current backend controllers.
    Authentication: all endpoints require Bearer token except /user/login.
servers:
  - url: http://localhost:8090
...
paths:
...
  /warehouse/stock/listPageC1:
    post:
      tags: [Warehouse]
      summary: Stock page query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryPageParamWarehouseStock"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/stock/detail:
    get:
      tags: [Warehouse]
      summary: Stock detail
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/stock/save:
    post:
      tags: [Warehouse]
      summary: Create stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseStock"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/stock/update:
    post:
      tags: [Warehouse]
      summary: Update stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseStock"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/inbound:
    post:
      tags: [Warehouse]
      summary: Manual inbound (apply directly to stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockIoRequest"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/inbound/apply:
    post:
      tags: [Warehouse]
      summary: Submit inbound approval flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockIoRequest"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/outbound:
    post:
      tags: [Warehouse]
      summary: Manual outbound (apply directly to stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockIoRequest"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/outbound/apply:
    post:
      tags: [Warehouse]
      summary: Submit outbound approval flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockIoRequest"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/flow/listPageC1:
    post:
      tags: [Warehouse]
      summary: Warehouse flow page query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryPageParamWarehouseFlow"
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/flow/approve:
    post:
      tags: [Warehouse]
      summary: Approve warehouse flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
            example:
              id: 301
              operatorId: 1
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /warehouse/flow/reject:
    post:
      tags: [Warehouse]
      summary: Reject warehouse flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
            example:
              id: 301
              operatorId: 1
      responses:
        "200":
          description: Result wrapper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Result:
      type: object
      description: Loose generic response wrapper for compatibility in Apifox validation.
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        total:
          type: integer
          format: int64
          example: 0
        data:
          description: Generic business payload, shape varies by endpoint and may be null.
          example: null
...
    QueryPageParamWarehouseStock:
      type: object
      properties:
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        param:
          type: object
          properties:
            keyword:
              type: string
    QueryPageParamWarehouseFlow:
      type: object
      properties:
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        param:
          type: object
          properties:
            productCode:
              type: string
            flowType:
              type: string
              description: for example pin/pout/in/out
    ApproveRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        operatorId:
          type: integer
      required: [id, operatorId]
...
    StockIoRequest:
      type: object
      properties:
        productCode:
          type: string
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        bizDate:
          type: string
          format: date
        remark:
          type: string
        operatorId:
          type: integer
      required: [productCode, productName, quantity]
    WarehouseStock:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productCode:
          type: string
        productName:
          type: string
        specification:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minStock:
          type: number
        warehouseLocation:
          type: string
        remark:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
security:
  - bearerAuth: []